🐛(frontend) fix issue when no template
A bug occured when no template was
available. This commit fixes this issue.
